add Title Strategy All clear 2, 5, 10

+++ src/test/java/com/anzymus/neogeo/hiscores/success/AbstractAllClearerTitleStrategyTest.java
@@ -14,41 +14,40 @@ import com.anzymus.neogeo.hiscores.service.TitleService;
 
 public abstract class AbstractAllClearerTitleStrategyTest {
 
-	TitleUnlockingStrategy titleUnlockingStrategy = getAllClearStrategy();
+    TitleUnlockingStrategy titleUnlockingStrategy = getAllClearStrategy();
 
-	@Mock
-	TitleService titleService;
+    @Mock
+    TitleService titleService;
 
-	Player player;
+    Player player;
 
-	@Before
-	public void init() {
-		MockitoAnnotations.initMocks(this);
-		titleUnlockingStrategy.initialize(titleService);
-		player = new Player();
-	}
+    @Before
+    public void init() {
+        MockitoAnnotations.initMocks(this);
+        titleUnlockingStrategy.initialize(titleService);
+        player = new Player();
+    }
 
-	@Test
-	public void should_detect_unlocking() {
-		when(titleService.getNumAllClearsByPlayer(player)).thenReturn(
-				getNumAllClearsToCreate());
+    @Test
+    public void should_detect_unlocking() {
+        when(titleService.getNumAllClearsByPlayer(player)).thenReturn(getNumAllClearsToCreate());
 
-		boolean unlocked = titleUnlockingStrategy.isUnlocked(player);
+        boolean unlocked = titleUnlockingStrategy.isUnlocked(player);
 
-		assertTrue(unlocked);
-	}
+        assertTrue(unlocked);
+    }
 
-	@Test
-	public void should_not_detect_unlocking() {
-		when(titleService.getNumAllClearsByPlayer(player)).thenReturn(0L);
+    @Test
+    public void should_not_detect_unlocking() {
+        when(titleService.getNumAllClearsByPlayer(player)).thenReturn(0);
 
-		boolean unlocked = titleUnlockingStrategy.isUnlocked(player);
+        boolean unlocked = titleUnlockingStrategy.isUnlocked(player);
 
-		assertFalse(unlocked);
-	}
+        assertFalse(unlocked);
+    }
 
-	protected abstract long getNumAllClearsToCreate();
+    protected abstract int getNumAllClearsToCreate();
 
-	protected abstract TitleUnlockingStrategy getAllClearStrategy();
+    protected abstract TitleUnlockingStrategy getAllClearStrategy();
 
 }
